+++ extensions/twitter/src/main/java/app/revanced/twitter/patches/links/ChangeLinkSharingDomainPatch.java
@@ -2,13 +2,23 @@ package app.revanced.twitter.patches.links;
 
 @SuppressWarnings("unused")
 public final class ChangeLinkSharingDomainPatch {
-    private static final String LINK_FORMAT = "%s/%s/status/%s";
+    private static final String LINK_FORMAT = "https://%s/%s/status/%s";
 
     /**
      * Method is modified during patching.  Do not change.
      */
     private static String getShareDomain() {
         return "";
+    }
+
+    // TODO remove this once changeLinkSharingDomainResourcePatch is restored
+    /**
+     * Injection point.
+     */
+    public static String formatResourceLink(Object... formatArgs) {
+        String username = (String) formatArgs[0];
+        String tweetId = (String) formatArgs[1];
+        return String.format(LINK_FORMAT, getShareDomain(), username, tweetId);
     }
 
     /**