+++ patches/src/main/kotlin/app/revanced/patches/music/misc/tracks/ForceOriginalAudioPatch.kt
@@ -3,13 +3,10 @@ package app.revanced.patches.music.misc.tracks
 import app.revanced.patches.music.misc.extension.sharedExtensionPatch
 import app.revanced.patches.music.misc.settings.PreferenceScreen
 import app.revanced.patches.music.misc.settings.settingsPatch
-import app.revanced.patches.music.playservice.is_8_10_or_greater
+import app.revanced.patches.music.playservice.is_8_05_or_greater
 import app.revanced.patches.music.playservice.versionCheckPatch
 import app.revanced.patches.music.shared.mainActivityOnCreateFingerprint
 import app.revanced.patches.shared.misc.audio.forceOriginalAudioPatch
-
-private const val EXTENSION_CLASS_DESCRIPTOR =
-    "Lapp/revanced/extension/music/patches/ForceOriginalAudioPatch;"
 
 @Suppress("unused")
 val forceOriginalAudioPatch = forceOriginalAudioPatch(
@@ -27,8 +24,8 @@ val forceOriginalAudioPatch = forceOriginalAudioPatch(
             )
         )
     },
-    fixUseLocalizedAudioTrackFlag = { is_8_10_or_greater },
+    fixUseLocalizedAudioTrackFlag = { is_8_05_or_greater },
     mainActivityOnCreateFingerprint = mainActivityOnCreateFingerprint,
-    subclassExtensionClassDescriptor = EXTENSION_CLASS_DESCRIPTOR,
+    subclassExtensionClassDescriptor = "Lapp/revanced/extension/music/patches/ForceOriginalAudioPatch;",
     preferenceScreen = PreferenceScreen.MISC,
 )